fix(encoder-patch): single source of truth for Qwen3-Omni benchmark setup
Address PR #436 review 4295981945 and the same divergence-risk pattern
in two adjacent helpers.

The session-scoped qwen3_omni_vision_sglang_env fixture was hard-coding
ServerArgs.model_path to the HF hub name while cuda_device and the
test-local AutoConfig.from_pretrained calls honored
SGLANG_OMNI_TEST_QWEN3_MODEL. An offline runner setting the env var to
a local checkpoint would have SGLang initialize global state from the
hub name and fail before the parity / perf assertions, and even when
the remote cache existed the SGLang runtime path and the weight-load
path could silently diverge.

Hoist a single QWEN3_OMNI_TEST_MODEL_PATH constant in conftest.py that
resolves the env var once, and reuse it from both ServerArgs.model_path
and the two benchmark modules so the HF path and the SGLang runtime
path cannot drift again. While there, lift the byte-identical
_model_cache_present helper, _resolve_model_dir helper, and the module
cuda_device fixture into the same conftest so a future cache-resolution
change to one module cannot silently diverge from the other.

Verified on the H100 dev container with the combined command
`pytest -v -s -m benchmark
  tests/test_model/test_v1_encoder_qwen3_vl_hf_parity.py
  tests/test_model/test_v1_encoder_patches_perf.py`:
5 passed in 2088.51s (parity cos>=0.9997, perf prep_ratio 0.0007%-0.0323%).
