+++ tests/test_model/test_v1_encoder_patches_perf.py
@@ -21,12 +21,12 @@ local HF cache (skips if absent — never auto-downloads).
 from __future__ import annotations
 
 import json
-from pathlib import Path
 
 import pytest
 import torch
 
 from tests.test_model.conftest import QWEN3_OMNI_TEST_MODEL_PATH as MODEL_PATH
+from tests.test_model.conftest import resolve_qwen3_omni_model_dir
 
 # (label, t, h, w, n_grids, n_iters)
 SCENARIOS = [
@@ -43,44 +43,6 @@ SCENARIOS = [
 # This 5% gate only fires on a real regression; tighter trend tracking
 # happens via the per-scenario PERF_JSON CI artifacts instead.
 PREP_RATIO_GATE = 0.05
-
-
-def _model_cache_present(model_path: str) -> bool:
-    try:
-        from huggingface_hub import snapshot_download
-    except ImportError:
-        return False
-
-    if Path(model_path).exists():
-        return True
-    try:
-        snapshot_download(model_path, local_files_only=True)
-    except Exception:
-        return False
-    return True
-
-
-def _resolve_model_dir(model_path: str) -> Path:
-    """Resolve to the model directory without triggering a download."""
-    if Path(model_path).exists():
-        return Path(model_path)
-    from huggingface_hub import snapshot_download
-
-    return Path(snapshot_download(model_path, local_files_only=True))
-
-
-@pytest.fixture(scope="module")
-def cuda_device():
-    if not torch.cuda.is_available():
-        pytest.skip("CUDA not available")
-    if not _model_cache_present(MODEL_PATH):
-        pytest.skip(
-            f"{MODEL_PATH} is not in the local HF cache; this benchmark test "
-            f"refuses to auto-download a multi-GB checkpoint. Pre-populate "
-            f"the cache or set SGLANG_OMNI_TEST_QWEN3_MODEL to a local path."
-        )
-    torch.cuda.set_device(0)
-    return torch.device("cuda:0")
 
 
 @pytest.fixture(scope="module")
@@ -124,7 +86,7 @@ def sg_visual(cuda_device, qwen3_omni_vision_sglang_env):
 def _load_visual_safetensors():
     from safetensors import safe_open
 
-    md = _resolve_model_dir(MODEL_PATH)
+    md = resolve_qwen3_omni_model_dir(MODEL_PATH)
     weight_map = json.loads((md / "model.safetensors.index.json").read_text())[
         "weight_map"
     ]
